+++ src/index.css
@@ -1,22 +1,73 @@
 .wrapper {
-    margin: 50px auto;
-    padding: 50px;
+    margin: 100px auto;
+    /* padding: 50px; */
+    display: flex;
+    flex-direction: column;
+    max-width: fit-content;
+    background-color: #fff;
+    -webkit-box-shadow: 4px 4px 8px 0px rgba(34, 60, 80, 0.2);
+    -moz-box-shadow: 4px 4px 8px 0px rgba(34, 60, 80, 0.2);
+    box-shadow: 4px 4px 8px 0px rgba(34, 60, 80, 0.2);
+}
+
+.header,
+.footer {
+    position: relative;
+    display: flex;
+}
+
+.body {
     display: flex;
     gap: 20px;
-    /* flex-direction: column; */
     justify-content: center;
-    /* border: 1px solid red; */
-    max-width: fit-content;
+    background-color: var(--BackgroundDark);
+}
+
+.custom__loadFile,
+.custom__saveFile {
+    position: relative;
+    width: 100%;
+    height: 70px;
+    cursor: pointer;
+    background-color: #63caab50;
+}
+
+.custom-load__image,
+.custom-download__image {
+    position: absolute;
+    top: 50%;
+    left: 10%;
+    transform: translateY(-50%);
+}
+
+.custom-load__text,
+.custom-download__text {
+    position: absolute;
+    font-size: 30px;
+    font-weight: 200;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
 }
 
 #inputLoad {
-    margin: 20px auto;
-    width: fit-content;
-    /* border: 1px solid red; */
+    margin: 20px auto 0;
+    height: 50px;
+    visibility: hidden;
+    opacity: 0;
+}
+
+
+.custom__saveFile {
+    border: 1px solid red;
+    border: none;
+    font-weight: 200;
+    font-size: 24px;
+    cursor: pointer;
 }
 
 .canvas-wrapper {
-
+    height: 500px;
     position: relative;
 }
 
@@ -42,51 +93,51 @@
     overflow: hidden;
     font-size: 42px;
     font-weight: 900;
-    cursor: move;
+    cursor: default;
     /* border: 1px solid red; */
     color: #fff;
     letter-spacing: 6px;
-    text-shadow: 
-		-0   -4px 1px #000000,
-		 0   -4px 1px #000000,
-		-0    4px 1px #000000,
-		 0    4px 1px #000000,
-		-4px -0   1px #000000,
-		 4px -0   1px #000000,
-		-4px  0   1px #000000,
-		 4px  0   1px #000000,
-		-1px -4px 1px #000000,
-		 1px -4px 1px #000000,
-		-1px  4px 1px #000000,
-		 1px  4px 1px #000000,
-		-4px -1px 1px #000000,
-		 4px -1px 1px #000000,
-		-4px  1px 1px #000000,
-		 4px  1px 1px #000000,
-		-2px -4px 1px #000000,
-		 2px -4px 1px #000000,
-		-2px  4px 1px #000000,
-		 2px  4px 1px #000000,
-		-4px -2px 1px #000000,
-		 4px -2px 1px #000000,
-		-4px  2px 1px #000000,
-		 4px  2px 1px #000000,
-		-3px -4px 1px #000000,
-		 3px -4px 1px #000000,
-		-3px  4px 1px #000000,
-		 3px  4px 1px #000000,
-		-4px -3px 1px #000000,
-		 4px -3px 1px #000000,
-		-4px  3px 1px #000000,
-		 4px  3px 1px #000000,
-		-4px -4px 1px #000000,
-		 4px -4px 1px #000000,
-		-4px  4px 1px #000000,
-		 4px  4px 1px #000000,
-		-4px -4px 1px #000000,
-		 4px -4px 1px #000000,
-		-4px  4px 1px #000000,
-		 4px  4px 1px #000000;
+    text-shadow:
+        -0 -4px 1px #000000,
+        0 -4px 1px #000000,
+        -0 4px 1px #000000,
+        0 4px 1px #000000,
+        -4px -0 1px #000000,
+        4px -0 1px #000000,
+        -4px 0 1px #000000,
+        4px 0 1px #000000,
+        -1px -4px 1px #000000,
+        1px -4px 1px #000000,
+        -1px 4px 1px #000000,
+        1px 4px 1px #000000,
+        -4px -1px 1px #000000,
+        4px -1px 1px #000000,
+        -4px 1px 1px #000000,
+        4px 1px 1px #000000,
+        -2px -4px 1px #000000,
+        2px -4px 1px #000000,
+        -2px 4px 1px #000000,
+        2px 4px 1px #000000,
+        -4px -2px 1px #000000,
+        4px -2px 1px #000000,
+        -4px 2px 1px #000000,
+        4px 2px 1px #000000,
+        -3px -4px 1px #000000,
+        3px -4px 1px #000000,
+        -3px 4px 1px #000000,
+        3px 4px 1px #000000,
+        -4px -3px 1px #000000,
+        4px -3px 1px #000000,
+        -4px 3px 1px #000000,
+        4px 3px 1px #000000,
+        -4px -4px 1px #000000,
+        4px -4px 1px #000000,
+        -4px 4px 1px #000000,
+        4px 4px 1px #000000,
+        -4px -4px 1px #000000,
+        4px -4px 1px #000000,
+        -4px 4px 1px #000000,
+        4px 4px 1px #000000;
 }
 
 .upperText {
@@ -100,7 +151,6 @@
 .upperText__item,
 .lowerText__item {
     margin: 0 auto;
-    /* border: 1px solid red; */
 }
 
 #upperText {
@@ -118,9 +168,10 @@
 
 .nav {
     display: flex;
-    /* gap: 20px; */
     flex-direction: column;
-    border: 1px solid red;
+    height: 500px;
+
+    background-color: #fff;
 }
 
 .nav-label {
@@ -129,16 +180,25 @@
     gap: 20px;
     flex-direction: column;
     height: fit-content;
-    border: 1px solid red;
 }
 
+.input-text {
+    padding: 10px 20px;
+    width: 100%;
+    font-size: 24px;
+    border: none;
+    background-color: var(--BackgroundDark);
+}
 
+.nav-label>* {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
 
 #fontsize_up,
 #fontsize_lo {
-    margin: 0 auto;
-    width: 80px;
-    font-size: 24px;
-    font-weight: 900;
-
+    width: 70px;
+    font-size: 20px;
+    border: none;
 }
