+++ src/index.html
@@ -19,37 +19,57 @@
 
 <body>
 
-    <input type="file" id="inputLoad">
+
     <div class="wrapper">
-        <div class="canvas-wrapper">
-            <canvas id="canvas"></canvas>
-            <div class="upperText">
-                <span class="upperText__item"></span>
-            </div>
-            <div class="lowerText">
-                <span class="lowerText__item"></span>
-            </div>
-        </div>
-        <div class="nav">
-            <label class="nav-label">
-                <span>Enter upper text</span>
-                <input id="textInput_up" type="text" placeholder="Your text...">
-            </label>
-            <label class="nav-label">
-                <span>Font size from upper text</span>
-                <input id="fontsize_up" type="number" value="45">
-            </label>
-            <label class="nav-label">
-                <span>Enter lower text</span>
-                <input id="textInput_lo" type="text" placeholder="Your text...">
-            </label>
-            <label class="nav-label">
-                <span>Font size from lower text</span>
-                <input id="fontsize_lo" type="number" value="45">
+        <div class="header">
+            <label class="custom__loadFile">
+                <input type="file" id="inputLoad">
+                <img class="custom-load__image" src="./image/upload-image.png" width="50" height="50">
+                <span class="custom-load__text">Load image</span>
             </label>
         </div>
+        <div class="body">
+            <div class="canvas-wrapper">
+                <canvas id="canvas"></canvas>
+                <div class="upperText">
+                    <span class="upperText__item"></span>
+                </div>
+                <div class="lowerText">
+                    <span class="lowerText__item"></span>
+                </div>
+            </div>
+            <div class="nav">
+                <label class="nav-label">
+                    <span>Enter upper text</span>
+                    <input class="input-text" id="textInput_up" type="text" placeholder="Your text...">
+                </label>
+
+                <label class="nav-label">
+                    <span>Enter lower text</span>
+                    <input class="input-text" id="textInput_lo" type="text" placeholder="Your text...">
+                </label>
+                <label class="nav-label">
+                    <div>
+                        <span>Font size from upper text</span>
+                        <input id="fontsize_up" type="number" value="45">
+                    </div>
+                    <div>
+                        <span>Font size from lower text</span>
+                        <input id="fontsize_lo" type="number" value="45">
+                    </div>
+                </label>
+            </div>
+        </div>
+        <div class="footer">
+            <button class="custom__saveFile" id="buttonDownload">
+                <img class="custom-download__image" src="./image/download-image.png" width="50" height="50">
+                Download
+            </button>
+        </div>
+
     </div>
     <script type="module" src="./js/index.js"></script>
+    <script type="module" src="./js/text.js"></script>
 </body>
 
 </html>
